myenv/requirements.txt
python-dotenv==1.0.1
crewai
crewai_tools
langchain_groq==0.1.9
tenacity==9.0.0
protobuf==4.24.4